create helper function fto format ingredients. apply to Recipe Form and recipe page.
